@@ -29,7 +29,7 @@ namespace SeleniumExtras.MemberBuilders
     /// </summary>
     internal class WebElementBuilder : IMemberBuilder
     {
-        public bool CreateObject(Type memberType, IElementLocator locator, IEnumerable<By> bys, bool cache, out object createdObject)
+        public bool CreateObject(Type memberType, IElementLocator locator, IEnumerable<By> bys, bool cache, out object? createdObject)
         {
             createdObject = null;
 
