@@ -8,7 +8,7 @@ namespace SeleniumExtras.MemberBuilders
 {
     internal class WrappedElementListBuilder : IMemberBuilder
     {
-        public bool CreateObject(Type memberType, IElementLocator locator, IEnumerable<By> bys, bool cache, out object createdObject)
+        public bool CreateObject(Type memberType, IElementLocator locator, IEnumerable<By> bys, bool cache, out object? createdObject)
         {
             createdObject = null;
 
